ci: add eslint and prettier to the API
