--- api/config/db.js
@@ -1,17 +1,0 @@
-const mongoose = require("mongoose");
-
-const connectDB = async () => {
-  try {
-    const conn = await mongoose.connect(process.env.MONGO_URI, {
-      useNewUrlParser: true,
-      useCreateIndex: true,
-      useFindAndModify: false,
-      useUnifiedTopology: true
-    });
-    console.log(`MongoDB Connected: ${conn.connection.host}`);
-  } catch (err) {
-    console.log("Error db.js ", err);
-  }
-};
-
-module.exports = connectDB;
